Move icon-related functions to glue.icons
