+++ glue/app/qt/layer_tree_widget.py
@@ -12,7 +12,8 @@ from glue.config import single_subset_action
 from glue import core
 from glue.qt import qtutil
 from glue.dialogs.link_editor.qt import LinkEditor
-from glue.qt.qtutil import get_icon, action
+from glue.icons.qt import get_icon
+from glue.qt.qtutil import action
 from glue.dialogs.custom_component.qt import CustomComponentWidget
 from glue.dialogs.subset_facet.qt import SubsetFacet
 from glue.dialogs.data_wizard.qt import data_wizard
