+++ glue/app/qt/layer_tree_widget_ui.py
@@ -2,7 +2,8 @@ from __future__ import absolute_import, division, print_function
 
 from glue.external.qt import QtCore, QtGui, is_pyqt5
 from glue.core.qt.data_collection_model import DataCollectionView
-from glue.qt.qtutil import GlueActionButton, get_icon
+from glue.icons.qt import get_icon
+from glue.qt.qtutil import GlueActionButton
 
 
 class Ui_LayerTree(object):
