+++ glue/core/qt/data_collection_model.py
@@ -9,7 +9,8 @@ from glue.core import message as m
 from glue.core.decorators import memoize
 from glue import core
 from glue.core.qt.mime import LAYERS_MIME_TYPE
-from glue.qt.qtutil import layer_icon
+from glue.icons.qt import layer_icon
+
 from glue.core.qt.style_dialog import StyleDialog
 from glue.utils.qt import PyMimeData
