+++ glue/core/qt/layer_artist_model.py
@@ -15,7 +15,7 @@ from glue.external.qt.QtCore import Qt
 from glue.external.qt import QtGui, QtCore
 from glue.core.layer_artist import LayerArtistBase, LayerArtistContainer
 from glue.core.qt.style_dialog import StyleDialog
-from glue.qt.qtutil import layer_artist_icon
+from glue.icons.qt import layer_artist_icon
 from glue.core.qt.mime import LAYERS_MIME_TYPE
 from glue.utils import nonpartial
 from glue.utils.qt import PythonListModel, PyMimeData
