+++ glue/icons/qt/__init__.py
@@ -1,0 +1,1 @@
+from .helpers import *
