+++ glue/viewers/image/qt/viewer_widget.py
@@ -12,7 +12,7 @@ from glue.viewers.image.client import MplImageClient
 from glue.viewers.common.qt.toolbar import GlueToolbar
 from glue.viewers.common.qt.mouse_mode import (RectangleMode, CircleMode, PolyMode,
                                 ContrastMode)
-from glue.qt.qtutil import get_icon
+from glue.icons.qt import get_icon
 from glue.utils.qt.widget_properties import CurrentComboProperty, ButtonProperty, connect_current_combo, _find_combo_data
 from glue.viewers.common.qt.data_slice_widget import DataSlice
 from glue.viewers.common.qt.data_viewer import DataViewer
